+++ internal/mono/mono.go
@@ -1,0 +1,1 @@
+package mono
