+++ internal/mono/mono_test.go
@@ -1,0 +1,1 @@
+package mono_test
