+++ internal/rgbd/rgbd.go
@@ -1,0 +1,1 @@
+package rgbd
