+++ viam-orb-slam3_test.go
@@ -4,27 +4,20 @@
 package viamorbslam3_test
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"image"
-	"net"
 	"os"
 	"strconv"
 	"sync"
-	"sync/atomic"
 	"testing"
 
 	"github.com/edaniels/golog"
 	"github.com/edaniels/gostream"
 	"github.com/pkg/errors"
 	"go.viam.com/rdk/components/camera"
-	"go.viam.com/rdk/config"
-	"go.viam.com/rdk/pointcloud"
-	"go.viam.com/rdk/registry"
 	"go.viam.com/rdk/rimage"
 	"go.viam.com/rdk/rimage/transform"
-	"go.viam.com/rdk/services/slam"
 	"go.viam.com/rdk/testutils/inject"
 	rdkutils "go.viam.com/rdk/utils"
 	slamConfig "go.viam.com/slam/config"
@@ -32,9 +25,7 @@ import (
 	"go.viam.com/test"
 	"go.viam.com/utils"
 	"go.viam.com/utils/artifact"
-	"google.golang.org/grpc"
 
-	viamorbslam3 "github.com/viamrobotics/viam-orb-slam3"
 	"github.com/viamrobotics/viam-orb-slam3/internal/testhelper"
 )
 
@@ -54,388 +45,7 @@ var (
 	_false                            = false
 )
 
-func getNumOrbslamImages(mode viamorbslam3.SubAlgo) int {
-	switch mode {
-	case viamorbslam3.Mono:
-		return 15
-	case viamorbslam3.Rgbd:
-		return 29
-	default:
-		return 0
-	}
-}
 
-func closeOutSLAMService(t *testing.T, name string) {
-	t.Helper()
-
-	if name != "" {
-		err := slamTesthelper.ResetFolder(name)
-		test.That(t, err, test.ShouldBeNil)
-	}
-}
-
-func setupTestGRPCServer(tb testing.TB) (*grpc.Server, int) {
-	listener, err := net.Listen("tcp", ":0")
-	test.That(tb, err, test.ShouldBeNil)
-	grpcServer := grpc.NewServer()
-	go grpcServer.Serve(listener)
-
-	return grpcServer, listener.Addr().(*net.TCPAddr).Port
-}
-
-func getGoodOrMissingDistortionParamsCamera(projA transform.Projector) *inject.Camera {
-	cam := &inject.Camera{}
-	cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-		return gostream.NewEmbeddedVideoStreamFromReader(
-			gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-				return image.NewNRGBA(image.Rect(0, 0, 1024, 1024)), nil, nil
-			}),
-		), nil
-	}
-	cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-		return nil, errors.New("camera not lidar")
-	}
-	cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-		return projA, nil
-	}
-	return cam
-}
-
-func setupDeps(attr *slamConfig.AttrConfig) registry.Dependencies {
-	deps := make(registry.Dependencies)
-	var projA transform.Projector
-	intrinsicsA := &transform.PinholeCameraIntrinsics{ // not the real camera parameters -- fake for test
-		Width:  1280,
-		Height: 720,
-		Fx:     200,
-		Fy:     200,
-		Ppx:    640,
-		Ppy:    360,
-	}
-	distortionsA := &transform.BrownConrady{RadialK1: 0.001, RadialK2: 0.00004}
-	projA = intrinsicsA
-
-	var projRealSense transform.Projector
-	intrinsicsRealSense := &transform.PinholeCameraIntrinsics{
-		Width:  1280,
-		Height: 720,
-		Fx:     900.538,
-		Fy:     900.818,
-		Ppx:    648.934,
-		Ppy:    367.736,
-	}
-	distortionsRealSense := &transform.BrownConrady{
-		RadialK1:     0.158701,
-		RadialK2:     -0.485405,
-		RadialK3:     0.435342,
-		TangentialP1: -0.00143327,
-		TangentialP2: -0.000705919,
-	}
-	projRealSense = intrinsicsRealSense
-
-	var projWebcam transform.Projector
-	intrinsicsWebcam := &transform.PinholeCameraIntrinsics{
-		Width:  640,
-		Height: 480,
-		Fx:     939.2693584627577,
-		Fy:     940.2928257873841,
-		Ppx:    320.6075282958033,
-		Ppy:    239.14408757087756,
-	}
-	distortionsWebcam := &transform.BrownConrady{
-		RadialK1:     0.046535971648456166,
-		RadialK2:     0.8002516496932317,
-		RadialK3:     -5.408034254951954,
-		TangentialP1: -8.996658362365533e-06,
-		TangentialP2: -0.002828504714921335,
-	}
-	projWebcam = intrinsicsWebcam
-
-	for _, sensor := range attr.Sensors {
-		cam := &inject.Camera{}
-		switch sensor {
-		case "invalid_sensor_type":
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				return nil, errors.New("this device does not stream images")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return nil, transform.NewNoIntrinsicsError("")
-			}
-			deps[camera.Named(sensor)] = cam
-		case "good_camera":
-			cam = getGoodOrMissingDistortionParamsCamera(projA)
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsA, DistortionParams: distortionsA}, nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "missing_distortion_parameters_camera":
-			cam = getGoodOrMissingDistortionParamsCamera(projA)
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsA, DistortionParams: nil}, nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "missing_camera_properties":
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				return gostream.NewEmbeddedVideoStreamFromReader(
-					gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-						return image.NewNRGBA(image.Rect(0, 0, 1024, 1024)), nil, nil
-					}),
-				), nil
-			}
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return projA, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{}, errors.New("somehow couldn't get properties")
-			}
-			deps[camera.Named(sensor)] = cam
-		case "good_color_camera":
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return projA, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsA, DistortionParams: distortionsA}, nil
-			}
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				imgBytes, err := os.ReadFile(artifact.MustPath("rimage/board1.png"))
-				if err != nil {
-					return nil, err
-				}
-				lazy := rimage.NewLazyEncodedImage(imgBytes, rdkutils.MimeTypePNG)
-				return gostream.NewEmbeddedVideoStreamFromReader(
-					gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-						return lazy, func() {}, nil
-					}),
-				), nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "good_depth_camera":
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return nil, transform.NewNoIntrinsicsError("")
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{}, nil
-			}
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				imgBytes, err := os.ReadFile(artifact.MustPath("rimage/board1_gray.png"))
-				if err != nil {
-					return nil, err
-				}
-				lazy := rimage.NewLazyEncodedImage(imgBytes, rdkutils.MimeTypePNG)
-				return gostream.NewEmbeddedVideoStreamFromReader(
-					gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-						return lazy, func() {}, nil
-					}),
-				), nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "bad_camera_no_stream":
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				return nil, errors.New("bad_camera_no_stream")
-			}
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return projA, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsA, DistortionParams: distortionsA}, nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "bad_camera_intrinsics":
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				return gostream.NewEmbeddedVideoStreamFromReader(
-					gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-						return image.NewNRGBA(image.Rect(0, 0, 1024, 1024)), nil, nil
-					}),
-				), nil
-			}
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return &transform.PinholeCameraIntrinsics{}, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{
-					IntrinsicParams:  &transform.PinholeCameraIntrinsics{},
-					DistortionParams: &transform.BrownConrady{},
-				}, nil
-			}
-			deps[camera.Named(sensor)] = cam
-		case "orbslam_int_color_camera":
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return projRealSense, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsRealSense, DistortionParams: distortionsRealSense}, nil
-			}
-			var index uint64
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				defer func() {
-					orbslamIntSynchronizeCamerasChan <- 1
-				}()
-				// Ensure the StreamFunc functions for orbslam_int_color_camera and orbslam_int_depth_camera run under
-				// the lock so that they release images in the same call to getSimultaneousColorAndDepth().
-				orbslamIntCameraMutex.Lock()
-				select {
-				case <-orbslamIntCameraReleaseImagesChan:
-					i := atomic.AddUint64(&index, 1) - 1
-					if i >= uint64(getNumOrbslamImages(viamorbslam3.Rgbd)) {
-						return nil, errors.New("No more orbslam color images")
-					}
-					imgBytes, err := os.ReadFile(artifact.MustPath("slam/mock_camera_short/rgb/" + strconv.FormatUint(i, 10) + ".png"))
-					if err != nil {
-						return nil, err
-					}
-					lazy := rimage.NewLazyEncodedImage(imgBytes, rdkutils.MimeTypePNG)
-					return gostream.NewEmbeddedVideoStreamFromReader(
-						gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-							return lazy, func() {}, nil
-						}),
-					), nil
-				default:
-					return nil, errors.Errorf("Color camera not ready to return image %v", index)
-				}
-			}
-			deps[camera.Named(sensor)] = cam
-		case "orbslam_int_depth_camera":
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return nil, transform.NewNoIntrinsicsError("")
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{}, nil
-			}
-			var index uint64
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				defer func() {
-					orbslamIntCameraMutex.Unlock()
-				}()
-				// Ensure StreamFunc for orbslam_int_color_camera runs first, so that we lock orbslamIntCameraMutex before
-				// unlocking it
-				<-orbslamIntSynchronizeCamerasChan
-				select {
-				case <-orbslamIntCameraReleaseImagesChan:
-					i := atomic.AddUint64(&index, 1) - 1
-					if i >= uint64(getNumOrbslamImages(viamorbslam3.Rgbd)) {
-						return nil, errors.New("No more orbslam depth images")
-					}
-					imgBytes, err := os.ReadFile(artifact.MustPath("slam/mock_camera_short/depth/" + strconv.FormatUint(i, 10) + ".png"))
-					if err != nil {
-						return nil, err
-					}
-					lazy := rimage.NewLazyEncodedImage(imgBytes, rdkutils.MimeTypePNG)
-					return gostream.NewEmbeddedVideoStreamFromReader(
-						gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-							return lazy, func() {}, nil
-						}),
-					), nil
-				default:
-					return nil, errors.Errorf("Depth camera not ready to return image %v", index)
-				}
-			}
-			deps[camera.Named(sensor)] = cam
-		case "orbslam_int_webcam":
-			cam.NextPointCloudFunc = func(ctx context.Context) (pointcloud.PointCloud, error) {
-				return nil, errors.New("camera not lidar")
-			}
-			cam.ProjectorFunc = func(ctx context.Context) (transform.Projector, error) {
-				return projWebcam, nil
-			}
-			cam.PropertiesFunc = func(ctx context.Context) (camera.Properties, error) {
-				return camera.Properties{IntrinsicParams: intrinsicsWebcam, DistortionParams: distortionsWebcam}, nil
-			}
-			var index uint64
-			cam.StreamFunc = func(ctx context.Context, errHandlers ...gostream.ErrorHandler) (gostream.VideoStream, error) {
-				select {
-				case <-orbslamIntWebcamReleaseImageChan:
-					i := atomic.AddUint64(&index, 1) - 1
-					if i >= uint64(getNumOrbslamImages(viamorbslam3.Mono)) {
-						return nil, errors.New("No more orbslam webcam images")
-					}
-					imgBytes, err := os.ReadFile(artifact.MustPath("slam/mock_mono_camera/rgb/" + strconv.FormatUint(i, 10) + ".png"))
-					if err != nil {
-						return nil, err
-					}
-					img, _, err := image.Decode(bytes.NewReader(imgBytes))
-					if err != nil {
-						return nil, err
-					}
-					var ycbcrImg image.YCbCr
-					rimage.ImageToYCbCrForTesting(&ycbcrImg, img)
-					return gostream.NewEmbeddedVideoStreamFromReader(
-						gostream.VideoReaderFunc(func(ctx context.Context) (image.Image, func(), error) {
-							return &ycbcrImg, func() {}, nil
-						}),
-					), nil
-				default:
-					return nil, errors.Errorf("Webcam not ready to return image %v", index)
-				}
-			}
-			deps[camera.Named(sensor)] = cam
-		case "gibberish":
-			return deps
-		default:
-			continue
-		}
-	}
-	return deps
-}
-
-func createSLAMService(
-	t *testing.T,
-	attrCfg *slamConfig.AttrConfig,
-	logger golog.Logger,
-	bufferSLAMProcessLogs bool,
-	success bool,
-	executableName string,
-) (slam.Service, error) {
-	t.Helper()
-
-	ctx := context.Background()
-	cfgService := config.Service{Name: "test", Type: "slam", Model: viamorbslam3.Model}
-	cfgService.ConvertedAttributes = attrCfg
-
-	deps := setupDeps(attrCfg)
-
-	sensorDeps, err := attrCfg.Validate("path")
-	if err != nil {
-		return nil, err
-	}
-	test.That(t, sensorDeps, test.ShouldResemble, attrCfg.Sensors)
-
-	viamorbslam3.SetCameraValidationMaxTimeoutSecForTesting(1)
-	viamorbslam3.SetDialMaxTimeoutSecForTesting(2)
-
-	svc, err := viamorbslam3.New(ctx, deps, cfgService, logger, bufferSLAMProcessLogs, executableName)
-
-	if success {
-		if err != nil {
-			return nil, err
-		}
-		test.That(t, svc, test.ShouldNotBeNil)
-		return svc, nil
-	}
-
-	test.That(t, svc, test.ShouldBeNil)
-	return nil, err
-}
 
 func TestGeneralNew(t *testing.T) {
 	logger := golog.NewTestLogger(t)
@@ -443,7 +53,7 @@ func TestGeneralNew(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 
 	t.Run("New slam service with no camera", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		test.That(t, err, test.ShouldBeNil)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{},
@@ -488,7 +98,7 @@ func TestORBSLAMNew(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 
 	t.Run("New orbslamv3 service with good camera in slam mode rgbd", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{"good_color_camera", "good_depth_camera"},
 			ConfigParams:  map[string]string{"mode": "rgbd"},
@@ -522,7 +132,7 @@ func TestORBSLAMNew(t *testing.T) {
 	})
 
 	t.Run("New orbslamv3 service that errors due to missing distortion_parameters not being provided in config", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{"missing_distortion_parameters_camera"},
 			ConfigParams:  map[string]string{"mode": "mono"},
@@ -543,7 +153,7 @@ func TestORBSLAMNew(t *testing.T) {
 	})
 
 	t.Run("New orbslamv3 service that errors due to not being able to get camera properties", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{"missing_camera_properties"},
 			ConfigParams:  map[string]string{"mode": "mono"},
@@ -580,7 +190,7 @@ func TestORBSLAMNew(t *testing.T) {
 	})
 
 	t.Run("New orbslamv3 service with good camera in slam mode mono", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{"good_color_camera"},
 			ConfigParams:  map[string]string{"mode": "mono"},
@@ -650,7 +260,7 @@ func TestORBSLAMDataProcess(t *testing.T) {
 	name, err := slamTesthelper.CreateTempFolderArchitecture(logger)
 	test.That(t, err, test.ShouldBeNil)
 
-	grpcServer, port := setupTestGRPCServer(t)
+	grpcServer, port := testhelper.SetupTestGRPCServer(t)
 	attrCfg := &slamConfig.AttrConfig{
 		Sensors:       []string{"good_color_camera"},
 		ConfigParams:  map[string]string{"mode": "mono"},
@@ -726,7 +336,7 @@ func TestEndpointFailures(t *testing.T) {
 	name, err := slamTesthelper.CreateTempFolderArchitecture(logger)
 	test.That(t, err, test.ShouldBeNil)
 
-	grpcServer, port := setupTestGRPCServer(t)
+	grpcServer, port := testhelper.SetupTestGRPCServer(t)
 	attrCfg := &slamConfig.AttrConfig{
 		Sensors:       []string{"good_color_camera"},
 		ConfigParams:  map[string]string{"mode": "mono", "test_param": "viam"},
@@ -772,7 +382,7 @@ func TestSLAMProcessSuccess(t *testing.T) {
 	test.That(t, err, test.ShouldBeNil)
 
 	t.Run("Test online SLAM process with default parameters", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{"good_color_camera", "good_depth_camera"},
 			ConfigParams:  map[string]string{"mode": "rgbd", "test_param": "viam"},
@@ -813,7 +423,7 @@ func TestSLAMProcessSuccess(t *testing.T) {
 	})
 
 	t.Run("Test offline SLAM process with default parameters", func(t *testing.T) {
-		grpcServer, port := setupTestGRPCServer(t)
+		grpcServer, port := testhelper.SetupTestGRPCServer(t)
 		attrCfg := &slamConfig.AttrConfig{
 			Sensors:       []string{},
 			ConfigParams:  map[string]string{"mode": "mono", "test_param": "viam"},
@@ -861,7 +471,7 @@ func TestSLAMProcessFail(t *testing.T) {
 	name, err := slamTesthelper.CreateTempFolderArchitecture(logger)
 	test.That(t, err, test.ShouldBeNil)
 
-	grpcServer, port := setupTestGRPCServer(t)
+	grpcServer, port := testhelper.SetupTestGRPCServer(t)
 	attrCfg := &slamConfig.AttrConfig{
 		Sensors:       []string{"good_color_camera"},
 		ConfigParams:  map[string]string{"mode": "mono", "test_param": "viam"},
